Fix TOA-94 Channel names for services with the same simple name result in conflicts
Fix Info:
* When two services result in the same channel name being used on the
same message bus Hornet should fail during messaging configuration to
prevent messages from the two services flowing over the same channel and
reaching unintended recipients. 
* Introduced nv.toa.failonchannelnamecollision configuration property
which can be used avoid throwing an exception if there is a collision to
allow for cases where this is intentional and to preserve backwards
compatibility. 

Also Fixed a bug in which the configuration property
nv.toa.generichandlerjoinsall was not being honored and upgraded the
configuration in test cases to use more up to date config DDL.